requirements.txt
PySide6
flatten-json
Pillow
exifread
psutil
watchdog